- Reviewed the structure for usePoint.
