+++ home-helper-front/pages/homeHelper/usePoint.tsx
@@ -1,30 +1,21 @@
-import {UsePointContainer} from "../../components/aggregations/use_point/UsePointContainer";
+import { Page } from "../../components/page/use_point/UsePointPage";
 
-const index = (props: {
-    data: {
-        exchangeItems: any[],
-        currentPoint: number,
-    },
-}) => {
-    return <UsePointContainer
-        exchangeItems={props.data.exchangeItems}
-        currentPoint={props.data.currentPoint}
-    />
-}
-export default index
+const index = (props: any) => {
+  return <Page data={props.data} />;
+};
+export default index;
 
 export async function getServerSideProps() {
-    const exchangeItems = [
-        {id: 1, count: 1, point: 500, name: 'お菓子'},
-        {id: 2, count: 1, point: 1000, name: 'お小遣い'},
-        {id: 3, count: 2, point: 3000, name: 'プラレール(3000円)'},
-    ]
+  const exchangeItems = [
+    { id: 1, count: 1, point: 500, name: "お菓子" },
+    { id: 2, count: 1, point: 1000, name: "お小遣い" },
+    { id: 3, count: 2, point: 3000, name: "プラレール(3000円)" },
+  ];
 
-    const data = {
-        exchangeItems: exchangeItems,
-        currentPoint: 30000,
-    }
+  const data = {
+    exchangeItems: exchangeItems,
+    currentPoint: 30000,
+  };
 
-
-    return {props: {data}}
+  return { props: { data } };
 }
